+++ src/Subjects/SubjectFactoryInterface.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * TechDivision\Import\Plugins\PluginInterface
+ * TechDivision\Import\Subjects\SubjectFactoryInterface
  *
  * NOTICE OF LICENSE
  *
@@ -18,12 +18,12 @@
  * @link      http://www.techdivision.com
  */
 
-namespace TechDivision\Import\Plugins;
+namespace TechDivision\Import\Subjects;
 
-use TechDivision\Import\Configuration\PluginConfigurationInterface;
+use TechDivision\Import\Configuration\SubjectConfigurationInterface;
 
 /**
- * The interface for all plugins.
+ * The interface for all subject factory implementations.
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
  * @copyright 2016 TechDivision GmbH <info@techdivision.com>
@@ -31,23 +31,15 @@ use TechDivision\Import\Configuration\PluginConfigurationInterface;
  * @link      https://github.com/techdivision/import
  * @link      http://www.techdivision.com
  */
-interface PluginInterface
+interface SubjectFactoryInterface
 {
 
     /**
-     * Process the plugin functionality.
+     * Factory method to create new subject instance.
      *
-     * @return void
-     * @throws \Exception Is thrown, if the plugin can not be processed
+     * @param \TechDivision\Import\Configuration\SubjectConfigurationInterface $subjectConfiguration The subject configuration
+     *
+     * @return \TechDivision\Import\Subjects\SubjectInterface The subject instance
      */
-    public function process();
-
-    /**
-     *  Set's the plugin configuration instance.
-     *
-     * @param \TechDivision\Import\Configuration\PluginConfigurationInterface $pluginConfiguration The plugin configuration instance
-     *
-     * @return void
-     */
-    public function setPluginConfiguration(PluginConfigurationInterface $pluginConfiguration);
+    public function createSubject(SubjectConfigurationInterface $subjectConfiguration);
 }